backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-extensions==4.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
pillow==11.3.0
PyJWT==2.9.0
python-dotenv==1.1.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
tzdata==2025.2
